pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.3.1",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.12",
    "psycopg[binary]>=3.1.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tqdm>=4.67.3",
]
